requirements.txt
# Core LLM Framework
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2
langchain-anthropic==0.1.1

# Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1

# PDF Processing (NO PyMuPDF - it has compilation issues)
pdfplumber==0.10.3
pypdf==3.17.4

# Web Search
tavily-python==0.3.3

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# UI
streamlit==1.30.0

# Utilities
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
requests==2.31.0

# Development
pytest==7.4.4
black==23.12.1
ipykernel==6.29.0
jupyter==1.0.0
jupyterlab==4.1.0
